BDOG-691: Tests, and a slight restructuring
There is a slight disconnect in the design of this, as Bootstrap
contains the DefaultHttpClient, and this library does not depend on
Bootstrap. Therefore, to make it work, Guice'd applications would
currently have to define this module:

    import play.api.{Configuration, Environment}
    import play.api.inject.{Binding, Module}
    import uk.gov.hmrc.http.CoreGet
    import uk.gov.hmrc.play.bootstrap.http.DefaultHttpClient

    class CookieModule extends Module {
      override def bindings(environment: Environment, configuration: Configuration): Seq[Binding[_]] = Seq(
        bind[CoreGet].to[DefaultHttpClient]
      )
    }

CoreGet exists in http-verbs, so I had based the interface from that.
However, the actual implementation of this trait is
`DefaultHttpClient`...

I imagine there'll be another PR for this library soon, once we design
around this issue.