WIP - Multi-module with less code
Trying to compile the play25 stuff using play26 sources.
Configuration is different between the two play versions. Play25 does
not have config.getOptional[A](path = "")

Cross-compilation in other projects seemed to be able to provide a small
section of code for specific play versions, and it was mixed in
(somehow - not really sure how that worked).

Briefly tried having a common module, but that would require writing
lots of adapter code as common wouldn't depend on anything.